# api-specification.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "RS AWS Cloud Developer Shop API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://f6yrfp0qz1.execute-api.eu-central-1.amazonaws.com"
        }
    ],
    "paths": {
        "/products": {
            "get": {
                "summary": "Get all products",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "List of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductWithStock"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Add a new product",
                "operationId": "addProduct",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddProduct"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Product created"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "summary": "Get product by id",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    }
                }
            },
            "ProductWithStock": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "count": {
                        "type": "integer"
                    }
                }
            },
            "AddProduct": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    }
                }
            }
        }
    }
}